+++ NEWS.md
@@ -1,5 +1,8 @@
 # bvarnet 1.0.3.9000
 * Minor bugfix that let to an error when running a model without covariates.
+* `extract_draws()` now accepts several parameters at once.
+* Minor documentation fixes
+* Bugfix: Removing order constrained from adjacent category model.
 
 # bvarnet 1.0.2